requirements.txt
requests>=2.25.0
mysql-connector-python>=8.0.0
pandas>=1.3.0
schedule>=1.0.0
pywin32>=300; platform_system=="Windows"
python-binance>=1.0.16
ccxt>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
nltk>=3.6.0
textblob>=0.15.3
python-dotenv>=0.19.0
websocket-client>=1.2.0
ta>=0.10.0  # Technical Analysis library
